requirements.txt
tqdm
requests
PyPDF2
psycopg2-binary
python-dotenv
